host_pc/requirements.txt
pyserial
pyqtgraph
pyqt6
numpy
sounddevice